+++ scraper/src/main/kotlin/com/nixops/scraper/mapper/ModuleMapper.kt
@@ -1,0 +1,54 @@
+package com.nixops.scraper.mapper
+
+import com.nixops.openapi.model.ModuleCourses
+import com.nixops.scraper.model.Course
+import com.nixops.scraper.model.Module
+import org.springframework.stereotype.Service
+
+@Service
+class ModuleMapper(private val courseMapper: CourseMapper) {
+  fun moduleToApiModule(
+      module: Module,
+      courses: Set<Course> = setOf()
+  ): com.nixops.openapi.model.Module {
+    val lectures = mutableListOf<com.nixops.openapi.model.Course>()
+    val tutorials = mutableListOf<com.nixops.openapi.model.Course>()
+    val other = mutableListOf<com.nixops.openapi.model.Course>()
+
+    for (course in courses) {
+      val apiCourse = courseMapper.courseToApiCourse(course)
+
+      when (course.activityName) {
+        "Vorlesung",
+        "Vorlesung mit integrierten Übungen" -> {
+          lectures.add(apiCourse)
+        }
+        "Tutorium",
+        "Übung" -> {
+          tutorials.add(apiCourse)
+        }
+        else -> {
+          other.add(apiCourse)
+        }
+      }
+    }
+
+    val moduleCourse = ModuleCourses(lectures = lectures, tutorials = tutorials, other = other)
+
+    return com.nixops.openapi.model.Module(
+        id = module.id.value.toString(),
+        code = module.moduleCode,
+        title = module.moduleTitle,
+        titleEn = module.moduleTitleEn,
+        content = module.moduleContent,
+        contentEn = module.moduleContentEn,
+        outcome = module.moduleOutcome,
+        outcomeEn = module.moduleOutcomeEn,
+        methods = module.moduleMethods,
+        methodsEn = module.moduleMethodsEn,
+        exam = module.moduleExam,
+        examEn = module.moduleExamEn,
+        credits = module.moduleCredits,
+        courses = moduleCourse)
+  }
+}
